Add required options to Pool instantiation
We were creating a Pool object with no options. This defaults to a host
of 127.0.0.1, which is fine unless your DB is on another host (as when
running in a container), so we now populate the 'host' option with and
ENV var. We also pass 'user' and 'database' options as they are also
required.